Move code generation into its own module
To make it possible to implement customised code generators out of tree,
the rendering engine must become defined by a trait and file system walker
should use an instance of this trait to render the templates.

The rendering code get moved from Template.write_rust to Engine and
Templates stays only as a container to hold the parsed data.

The new walker traverse_dir() captures the content of a customisable
prelude file in each directory level and prepends the collected contents
to the template file.