template: Add @@ for raw code
Sometimes it's necessary to have some code in the view to do some
calculations. To support this, everything after @@ until newline gets
written to the generated code file.

This makes it possible to circumvent the shortcomings of the current
parser, while to writer has to take care that the resulting code is valid.